Update room_assignment.txt
Made some textual changes to the first few paragraphs.

I would also like to see us add the optional section if the room has more than 5 people in it.... that can sometime happen when we shuffle around staff pre-event.   We don't have a clean way to show people that on Load-in days they are in a certain room, then on Core_nights they are in a separate room.